packages/app/src/containers/molecules/SignUpFailureMessage/index.test.tsx
import * as ReactTesting from '@testing-library/react';
import React from 'react';
import { Provider } from 'react-redux';

import * as T from '^/types';

import {
  DDMMockStore,
  commonAfterEach,
  createDDMMockStore,
} from '^/utilities/test-util';

import SignUpFailureMessage, {
  DispatchProps,
  OwnProps,
  mapDispatchToProps,
} from './';

describe('DispatchProps', () => {
  let dispatch: jest.Mock;
  let dispatchProps: DispatchProps;

  beforeEach(() => {
    dispatch = jest.fn();
    dispatchProps = mapDispatchToProps(dispatch);
  });

  it('should dispatch an action when sign up api reset is requested', () => {
    expect(dispatch).toHaveBeenCalledTimes(0);
    dispatchProps.resetSignUpAPI();
    expect(dispatch).toHaveBeenCalledTimes(1);
  });
});

describe('Connected SignUpFailureMessage', () => {
  const props: OwnProps = {};
  let store: DDMMockStore;

  beforeEach(() => {
    store = createDDMMockStore(T.Language.KO_KR);
  });

  afterEach(ReactTesting.cleanup);
  afterEach(commonAfterEach);

  it('should not emit an error during rendering', () => {
    expect(() => {
      ReactTesting.render(
        <Provider store={store}>
          <SignUpFailureMessage {...props} />
        </Provider>,
      );
    }).not.toThrowError();
  });
});
